Created cookies function and sum page which contains user likes
