+++ package.json
@@ -12,8 +12,10 @@
     "@types/react": "^16.9.36",
     "dotenv": "^8.2.0",
     "dotenv-cli": "^3.1.0",
+    "js-cookie": "^2.2.1",
     "ley": "^0.5.0",
     "next": "9.4.4",
+    "next-cookies": "^2.0.3",
     "postgres": "^1.0.2",
     "react": "16.13.1",
     "react-dom": "16.13.1"
