+++ pages/contact.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import Head from 'next/head';
+import Header from '../components/Header';
+import Footer from '../components/Footer';
+
+export default function Contact() {
+  return (
+    <div>
+      <Head>
+        <title>Find-your-dog</title>
+        <link rel="icon" href="/favicon.jpg" />
+      </Head>
+      <Header />
+      <h1>contact</h1>
+      <Footer />
+    </div>
+  );
+}
